Added more logging, cleaned up duplicated code, added benchmark pipeline test file
