+++ mcs_benchmark_data/cli/templates/benchmark_name_benchmark_transformer.py.template
