+++ mcs_benchmark_data/cli/templates/benchmark_name_pipeline_test.py.template
