+++ mcs_benchmark_data/cli/templates/metadata.json.template
