+++ mcs_benchmark_data/cli/templates/submission_name_benchmark_name_submission_transformer.py.template
