+++ mcs_benchmark_data/cli/templates/submissions_metadata.jsonl.template
